Moved the Plain form action into the select plan file
